backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.7.2
pandas==2.1.4
numpy==1.26.3
faiss-cpu==1.7.4
openai>=1.12.0
xgboost==2.0.3
sentence-transformers==3.0.1
google-generativeai>=0.7.0
firebase-admin==6.2.0
python-dotenv==1.0.0
tf-keras
groq
